Move To Arguments
This changes the architecture to accept arguments and not environment
variables.  This is so we can directly pass through commands to tco to
actually run tests.  We can now use the library version of argparse.

I've also allowed the cluster to not be deleted properly by making this
a cli switch.